+++ markdowns/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Moved files from data-global/migrations to data/migrations. ([Tryller](https://github.com/jprzimba))
 - New configurations in `config.lua`: `timeToRegenMinuteStamina` and `timeToRegenMinutePremiumStamina`. ([Tryller](https://github.com/jprzimba))
 - Added a new attribute `attackspeed` to mounts.xml, allowing mounts to increase player attack speed. ([Tryller](https://github.com/jprzimba))
+- New Feature: `fairFightTimeRange` is now configurable in `config.lua`. ([Tryller](https://github.com/jprzimba))
 
 ## Added files